services/user_service/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
pyodbc
python-multipart
bcrypt==4.0.1